ReadMe changes
Added more info to the README.md file.
Here it is:

The basic gist of this game is: It's the game that always tries to strive for originality, but fails. But will it ever be original? Find out in the game!

This is a Clickteam Fusion game, so the source code is technically the .mfa file, so I'll include the .mfa file in this repository as well as the .json file.

WARNING: OPENING THE .MFA FILE REQUIRES CLICKTEAM FUSION 2.5. IF YOU DO NOT OWN CLICKTEAM FUSION 2.5, I ADVISE YOU TO USE THE .JSON FILE. (YOU ALSO MAY BUY CLICKTEAM, BUT THAT IS MUCH MORE COSTLY THAN JUST USING THE .JSON FILE.) With that said, this is the official Nobody's Done This, Have They? Source Code!

Download the game at:

https://gamejolt.com/games/ndtht/644953

https://adentheboi.itch.io/nobodys-done-this-have-they